feat: make localStorage to carts state hook
